feat: split commit-to-list to its own file
Split commitToList to its own file will allow libraries
to create their own custom format
